azaz
azaza

+++ game/src/screen_gameplay.cpp
@@ -133,7 +133,7 @@ void DrawGameplayScreen(const World* world, Shader shader)
                 Rectangle frameRec = {0.0f,0.0f,(float)world->texture[3].width, (float)world->texture[3].height};
                 // Vector2 tiger_origin = TransitionCoordinate(world->notes.notes[i]->sita,world->notes.notes[i]->r+EARTH_RADIUS);
                 Rectangle destRec = { EARTH_POSX, EARTH_POSY, (float)world->texture[3].width/6, (float)world->texture[3].height };
-                DrawTexturePro(world->texture[3], frameRec, destRec, (Vector2){(float)world->texture[3].width/2,world->notes.notes[i]->r+EARTH_RADIUS}, world->notes.notes[i]->sita*DEG2RAD,WHITE);
+                DrawTexturePro(world->texture[3], frameRec, destRec, (Vector2){(float)world->texture[3].width/2,(float)(world->notes.notes[i]->r+EARTH_RADIUS)}, world->notes.notes[i]->sita*DEG2RAD,WHITE);
             }
